# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards API
  description: API для управления банковскими картами и пользователями
  version: 1.0.0
  contact:
    name: Александр
    email: admin@bank.com
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/admin/users:
    get:
      tags: [Admin Users]
      summary: Получить всех пользователей с картами
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminUserDto'
        '403':
          description: Недостаточно прав (требуется ADMIN)

    post:
      tags: [Admin Users]
      summary: Создать нового пользователя
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Ошибка валидации
        '403':
          description: Недостаточно прав
  /api/admin/users/{id}:
    put:
      tags: [Admin Users]
      summary: Обновить пользователя
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Пользователь обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: Пользователь не найден

    delete:
      tags: [Admin Users]
      summary: Удалить пользователя
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Пользователь удалён
        '404':
          description: Пользователь не найден

  # ==================== ADMIN CARDS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
    AdminUserDto:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        cardIds:
          type: array
          items:
            type: integer
    CreateUserRequest:
      type: object
      required: [username, email, password, role]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
...